#103 距離の共通化
